fix: Fixed and change items based on review
Completed the separation of actions and reducers, and fixed the 404 routing when no page matches the elements in the page. Added comments count to post view. And fixed issue with Categories link error.
