+++ src/actions/types.js
@@ -1,0 +1,13 @@
+export const POSTS_FILTER = 'POSTS_FILTER';
+export const SORT_POSTS = 'SORT_POSTS';
+export const LOAD_POSTS = 'LOAD_POSTS';
+export const ADD_POST = 'ADD_POST';
+export const EDIT_POST = 'EDIT_POST';
+export const DELETE_POST = 'DELETE_POST';
+export const VOTE_POST = 'VOTE_POST';
+export const ADD_COMMENT = 'ADD_COMMENT';
+export const EDIT_COMMENT = 'EDIT_COMMENT';
+export const DELETE_COMMENT = 'DELETE_COMMENT';
+export const VOTE_COMMENT = 'VOTE_COMMENT';
+export const LOAD_CATEGORIES = 'LOAD_CATEGORIES';
+export const LOAD_POST_ITEM = 'LOAD_POST_ITEM';
